requirements.txt
flask==3.0.2
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0
mypy==1.8.0
psutil==5.9.8
schedule==1.2.1
SQLAlchemy==2.0.27
alembic==1.13.1
flask-restx==1.3.0
flask-cors==4.0.0 